Added documentation headers broke up HigherOrderSpecBuilderExtensions.cs into separate extension classes
