+++ Karlssberg.Motiv/And/AndBooleanResultDescription.cs
@@ -19,7 +19,7 @@ internal sealed class AndBooleanResultDescription<TMetadata>(
 
 
     public override IEnumerable<string> GetDetailsAsLines() =>
-        causalResults.GetBinaryDetailsAsLines(operationName);
+        causalResults.GetBinaryJustificationAsLines(operationName);
     
     private string ExplainReasons(BooleanResultBase<TMetadata> result)
     {
