+++ Karlssberg.Motiv/And/AndSpecDescription.cs
@@ -14,7 +14,7 @@ internal sealed class AndSpecDescription<TModel, TMetadata>(
     public IEnumerable<string> GetDetailsAsLines()
     {
         IEnumerable<SpecBase<TModel, TMetadata>> specs = [left, right];
-        return specs.GetBinaryDetailsAsLines("AND");
+        return specs.GetBinaryJustificationAsLines("AND");
     }
 
     private string Summarize(SpecBase<TModel, TMetadata> operand)
