+++ Karlssberg.Motiv/AndAlso/AndAlsoBooleanResultDescription.cs
@@ -18,7 +18,7 @@ internal sealed class AndAlsoBooleanResultDescription<TMetadata>(
         };
 
     public override IEnumerable<string> GetDetailsAsLines() =>
-        causalResults.GetBinaryDetailsAsLines(operationName);
+        causalResults.GetBinaryJustificationAsLines(operationName);
     
     private string ExplainReasons(BooleanResultBase<TMetadata> result)
     {
