+++ Karlssberg.Motiv/Or/OrBooleanResultDescription.cs
@@ -19,7 +19,7 @@ internal sealed class OrBooleanResultDescription<TMetadata>(
         };
 
     public override IEnumerable<string> GetDetailsAsLines() =>
-        causalResults.GetBinaryDetailsAsLines(operationName);
+        causalResults.GetBinaryJustificationAsLines(operationName);
 
     private string ExplainReasons(BooleanResultBase<TMetadata> result)
     {
