+++ Karlssberg.Motiv/OrElse/OrElseBooleanResultDescription.cs
@@ -18,7 +18,7 @@ internal sealed class OrElseBooleanResultDescription<TMetadata>(
         };
 
     public override IEnumerable<string> GetDetailsAsLines() =>
-        causalResults.GetBinaryDetailsAsLines(operationName);
+        causalResults.GetBinaryJustificationAsLines(operationName);
     
     private string ExplainReasons(BooleanResultBase<TMetadata> result)
     {
