+++ Karlssberg.Motiv/OrElse/OrElseSpecDescription.cs
@@ -14,7 +14,7 @@ internal sealed class OrElseSpecDescription<TModel, TMetadata>(
     public IEnumerable<string> GetDetailsAsLines()
     {
         IEnumerable<SpecBase<TModel, TMetadata>> specs = [left, right];
-        return specs.GetBinaryDetailsAsLines("OR ELSE");
+        return specs.GetBinaryJustificationAsLines("OR ELSE");
     }
 
     private string Summarize(SpecBase<TModel, TMetadata> operand)
