+++ Karlssberg.Motiv/PropositionExtensions.cs
@@ -2,23 +2,6 @@
 
 internal static class PropositionExtensions
 {
-    private const int Spaces = 4;
-
-    internal static string IndentAfterFirstLine(this string text, int levels = 1)
-    {
-        var indentation = string.Join("", Enumerable.Repeat(" ", Spaces * levels));
-        return text.Replace("\n", $"\n{indentation}");
-    }
-
-    internal static string Indent(this string text, int levels = 1)
-    {
-        var indentation = string.Join("", Enumerable.Repeat(" ", Spaces * levels));
-        return $"{indentation}{text.Replace("\n", $"\n{indentation}")}";
-    }
-
-    internal static string JoinLines(this IEnumerable<string> textCollection) =>
-        string.Join(Environment.NewLine, textCollection);
-
     internal static string ToReason(
         this ISpecDescription specDescription,
         bool isSatisfied) =>
