+++ Karlssberg.Motiv/Tutorial/7.Compositions.md
@@ -46,8 +46,9 @@ isEvenAndPositive.IsSatisfiedBy(-2).Reason; // "-2 is not positive"
 ```
 
 ## Boolean composition
-The boolean operators are indispensable when composing propositions.
-When two propositions are combined using a binary boolean operator, then the result is a new proposition.
+The boolean operators (such as `&`, `!`, `|` etc.) are indispensable when composing propositions.
+When two propositions are combined using a binary boolean operator, then the result is a new proposition that will 
+harvest the assertions and metadata from the original propositions.
 ```csharp
 var isEven = 
     Spec.Build(new IsEvenProposition())
@@ -59,9 +60,13 @@ var isPositive =
 
 var isEvenAndPositive = isEven & isPositive;
     
-isEvenAndPositive.IsSatisfiedBy(2).Reason;  // "is even & is positive"
-isEvenAndPositive.IsSatisfiedBy(-3).Reason; // "not even & not positive"
-isEvenAndPositive.IsSatisfiedBy(-2).Reason; // "not positive"
+var result = isEvenAndPositive.IsSatisfiedBy(2);
+
+result.Reason;         // "is even & is positive"
+result.Assertions;     // ["is even", "is positive"]
+result.Justifications; // AND
+                       //     is even
+                       //     is positive
 
 ```
 | [Back - Boolean Results](./6.BooleanResults.md) | [Next - Operations](./8.Operations.md)  |
