+++ Karlssberg.Motiv/Tutorial/8.Operations.md
@@ -11,7 +11,8 @@ new IsEvenProposition() & new IsPositiveProposition();
 ```
 The `&` operator (or `And()` method) is used to logically combine two propositions into a single proposition
 that is only satisfied when both of the original propositions are satisfied.
-It will always evaluate both operands, regardless of the result of the left operand.
+
+Both operands are always evaluated, regardless of the result of the left operand.
 
 (see also: [The logical And operator](../And/And.md))
 
@@ -32,7 +33,8 @@ new IsEvenProposition() | new IsPositiveProposition();
 The `|` operator (or `Or()` method) is used to combine two propositions into a single proposition that is
 satisfied when either
 of the original propositions is satisfied.
-It will always evaluate both operands, regardless of the result of the left operand.
+
+Both operands are always evaluated, regardless of the result of the left operand.
 
 (see also: [The logical Or operator](../Or/Or.md))
